[refactor] helper 클래스 제거
